App divide into components
