+++ src/components/BottomNav.js
@@ -1,0 +1,27 @@
+import React from 'react';
+import BottomNavigation, {
+  BottomNavigationAction,
+} from 'material-ui/BottomNavigation';
+import RestoreIcon from '@material-ui/icons/Restore';
+import ExploreIcon from '@material-ui/icons/Explore';
+
+class BottomNav extends React.Component {
+  state = {
+    value: 0,
+  };
+  handleChange = (event, value) => {
+    this.setState({ value });
+  };
+
+  render() {
+    const { value } = this.state;
+    return (
+      <BottomNavigation value={value} onChange={this.handleChange} showLabels>
+        <BottomNavigationAction label="My Chats" icon={<RestoreIcon />} />
+        <BottomNavigationAction label="Explore" icon={<ExploreIcon />} />
+      </BottomNavigation>
+    );
+  }
+}
+
+export default BottomNav;
