+++ src/components/MenuButton.js
@@ -1,0 +1,54 @@
+import React from 'react';
+import IconButton from 'material-ui/IconButton';
+import Menu, { MenuItem } from 'material-ui/Menu';
+
+class MenuButton extends React.Component {
+  state = {
+    anchorEl: null,
+  };
+
+  handleMenu = event => {
+    this.setState({ anchorEl: event.currentTarget });
+  };
+  handleClose = () => {
+    this.setState({ anchorEl: null });
+  };
+  render() {
+    const { menuList, icon } = this.props;
+    const { anchorEl } = this.state;
+    const open = Boolean(anchorEl);
+    return (
+      <React.Fragment>
+        <IconButton
+          aria-owns={open ? 'menu-appbar' : null}
+          aria-haspopup="true"
+          onClick={this.handleMenu}
+          color="inherit"
+        >
+          {icon}
+        </IconButton>
+        <Menu
+          id="menu-appbar"
+          anchorEl={anchorEl}
+          anchorOrigin={{
+            vertical: 'top',
+            horizontal: 'right',
+          }}
+          transformOrigin={{
+            vertical: 'top',
+            horizontal: 'right',
+          }}
+          open={open}
+          onClose={this.handleClose}
+        >
+          {menuList.map(menuItem => (
+            <MenuItem key={menuItem.menuItem} onClick={this.handleClose}>
+              {menuItem.menuItem}
+            </MenuItem>
+          ))}
+        </Menu>
+      </React.Fragment>
+    );
+  }
+}
+export default MenuButton;
