+++ src/components/UserAvatar.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import Avatar from 'material-ui/Avatar';
+import titleInitials from '../utils/title-initials';
+
+const UserAvatar = ({ name }) => {
+  return <Avatar>{titleInitials(name)}</Avatar>;
+};
+
+export default UserAvatar;
